Refactored StableMatching.py to main.py
- Also updated READMEs in this regard
