+++ README.TXT
@@ -1,19 +1,19 @@
 # Advanced Algorithms Assignment 1
 
-Assignment 1 submission work for CS 6045/W01 - Advanced Algorithms.
+Assignment 1 submission work for CS 6045/W01 - Advanced Algorithms.  
 
-## Team Members
+## Team Members  
 - Michael Hayes (mhayes64@students.kennesaw.edu)
 - Cody Bijeaux (cbijeaux@students.kennesaw.edu)
 - Amrutvyasa Ramasamy (aramasam@students.kennesaw.edu)
 
 ## Run Program
 
-Ensure Python 3.12 or higher is installed.
+Ensure Python 3.12 or higher is installed.  
 
-Open a terminal window and change directories to this project. Enter the following command:
-`python .\StableMatching.py`
+Open a terminal window and change directories to this project. `main.py` is the main Stable Matching script. Enter the following command:  
+`python .\main.py`  
 
-Input files under the MatchingInput folder will be used to perform the stable matching. Outputs for all will be displayed in the console.
+Input files under the MatchingInput folder will be used to perform the stable matching. Outputs for all will be displayed in the console.  
 
 For review corresponding outputs for each input file are saved under the MatchingOutput directory.
